containers/metrics-agent/requirements.txt
prometheus_client==0.20.0
grpcio==1.63.0
protobuf==4.25.3
opentelemetry-proto==1.24.0
pynvml==11.5.0
